Modified GoalZone script for increased usability
The GoalZone script now allows for level designers to use sliders and drop down menus to determine the behavior of a goal zone. The goal zone script currently only supports sortable objects.

Known Issues:
Currently the tag manager has two values for the sorting zone, one for objects and one for the goal itself. This will be modified in a future update for increased user friendliness.