Updated the TagManager to be more intuitive
TagManager has been updated to use only a single ZoneTag. This is done to make it more friendly to the level designers. Now they may simply set whether an object is a goal or a sortable and set the zone tag so that they match. Matching objects with goal zones will gain extra happiness.