Goal Zone Update
## Goal Zone Update
- The goal zone is activated by a non-trigger collider.
- It now checks the current level of the incoming object (e.g. at the same level as the collider) for a tag manager.
 - If it is found, it updates the relevant local variable and adds the object to its list of tracked objects based on the instance ID of the transform of the object (again, at the same level as the collider that triggered the goal zone.)
 - If it is not found it begins searching __up__ the hierarchy for a tag manager.
  - It will stop searching when it finds a tag manager or when it has reached the top level of the hierarchy.

What this means in practical terms is that **the tag manager must be at the same level or higher than the object's collision** but colliders can be as nested as you'd like underneath it.

This was implemented to prevent bugs that would be caused by objects entering the goal zone while parented to another object such as Leggy.