Game manager title screen bug fix
The game manager now implements a callback from the scene manager when loading into a scene which checks to ensure it is destroyed if loaded into the title screen or the level select.
These checks are performed via build index!!! Thus, be cautious not to modify the title screen and level select from 0 and 1 in the index, respectively.

This also fixes a minor bug in the IK target fallback to prevent a null reference error.
This also updates the title screen and tutorial to touch up game feel (image fitting & transition speed, respectively).