Updated Level 01 and 02. Bugfix for game manager.
The game manager was not appropriately destroying old game managers when loading into a level due to checks that were not agnostic of hierarchy position.

Updated level 01 and 02 to have default audio handlers (for testing purposes) and event systems to allow for pausing.

Updated camera controller and rig controls to ensure an FMOD Studio Listener was always present when switching cameras.

Updated audio handler to prevent duplicate instances of music playing.